refactor: ChatbotExtractor and PhindSearch extractor
